--- src/main/java/com/example/ddingsroom/controller/AdminController.java
@@ -1,12 +1,0 @@
-package com.example.ddingsroom.controller;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-@ResponseBody
-@Controller
-public class AdminController {
-    @GetMapping("/admin")
-    public String adminP() {
-        return "admin Controller";
-    }
-}
