--- src/main/java/com/example/ddingsroom/dto/CodeSendDTO.java
@@ -1,12 +1,0 @@
-package com.example.ddingsroom.dto;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CodeSendDTO {
-    private String email;
-}
