--- src/main/java/com/example/ddingsroom/dto/CodeVerifyDTO.java
@@ -1,13 +1,0 @@
-package com.example.ddingsroom.dto;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class CodeVerifyDTO {
-    private String email;
-    private String code;
-}
