--- src/main/java/com/example/ddingsroom/dto/CustomUserDetails.java
@@ -1,47 +1,0 @@
-package com.example.ddingsroom.dto;
-import com.example.ddingsroom.entity.UserEntity;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import java.util.ArrayList;
-import java.util.Collection;
-public class CustomUserDetails implements UserDetails {
-    private final UserEntity userEntity;
-    public CustomUserDetails(UserEntity userEntity) {
-        this.userEntity = userEntity;
-    }
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collection = new ArrayList<>();
-        collection.add(new GrantedAuthority() {
-            @Override
-            public String getAuthority() {
-                return userEntity.getRole();
-            }
-        });
-        return collection;
-    }
-    @Override
-    public String getPassword() {
-        return userEntity.getPassword();
-    }
-    @Override
-    public String getUsername() {
-        return userEntity.getUsername();
-    }
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
