--- src/main/java/com/example/ddingsroom/dto/ModifyPasswordDTO.java
@@ -1,15 +1,0 @@
-package com.example.ddingsroom.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ModifyPasswordDTO {
-    private String email;
-    private String password;
-}
