--- src/main/java/com/example/ddingsroom/entity/RefreshEntity.java
@@ -1,22 +1,0 @@
-package com.example.ddingsroom.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class RefreshEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;
-    private String refresh;
-    private String expiration;
-}
