--- src/main/java/com/example/ddingsroom/entity/UserEntity.java
@@ -1,29 +1,0 @@
-package com.example.ddingsroom.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Data
-@Setter
-@Getter
-public class UserEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private String email;
-    private String username;
-    private String password;
-    private String age;
-    private String studentNumber;
-    private String role;
-    private String state;
-    private LocalDateTime registrationDate;
-}
