--- src/main/java/com/example/ddingsroom/jwt/JWTFilter.java
@@ -1,78 +1,0 @@
-package com.example.ddingsroom.jwt;
-
-import com.example.ddingsroom.dto.CustomUserDetails;
-import com.example.ddingsroom.entity.UserEntity;
-import io.jsonwebtoken.ExpiredJwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-public class JWTFilter extends OncePerRequestFilter {
-    private final JWTUtil jwtUtil;
-    public JWTFilter(JWTUtil jwtUtil) {
-        this.jwtUtil = jwtUtil;
-    }
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        System.out.println("JWTFilter: 필터 실행됨 (요청 URI: " + request.getRequestURI() + ")");
-        // 헤더에서 access키에 담긴 토큰을 꺼냄
-        String accessToken = request.getHeader("Authorization"); // 이거 원래 access였음 -> 영상 확인필요
-        System.out.println("Access Token: " + accessToken);
-        // 토큰이 없다면 다음 필터로 넘김
-        if (accessToken == null) {
-            System.out.println("JWTFilter: No access token found, skipping authentication");
-            filterChain.doFilter(request, response);
-            return;
-        }
-        // 이것도 확인 필요.
-        accessToken = accessToken.substring(7);
-        accessToken = accessToken.trim();
-
-        // 토큰 만료 여부 확인, 만료시 다음 필터로 넘기지 않음
-        try {
-            jwtUtil.isExpired(accessToken);
-        } catch (ExpiredJwtException e) {
-            System.out.println("JWTFilter: Token expired");
-            //response body
-            PrintWriter writer = response.getWriter();
-            writer.print("access token expired");
-            //response status code
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
-        }
-        // 토큰이 access인지 확인 (발급시 페이로드에 명시)
-        String category = jwtUtil.getCategory(accessToken);
-        if (!category.equals("access")) {
-            System.out.println("JWTFilter: Invalid access token");
-            //response body
-            PrintWriter writer = response.getWriter();
-            writer.print("invalid access token");
-            //response status code
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
-        }
-        // username, role 값을 획득
-        String username = jwtUtil.getUsername(accessToken);
-        String role = jwtUtil.getRole(accessToken);
-
-        System.out.println("JWTFilter: Setting authentication for " + username + " with role " + role);
-
-        UserEntity userEntity = new UserEntity();
-        userEntity.setUsername(username);
-        userEntity.setRole(role);
-        CustomUserDetails customUserDetails = new CustomUserDetails(userEntity);
-
-        Authentication authToken = new UsernamePasswordAuthenticationToken(customUserDetails, null, customUserDetails.getAuthorities());
-        SecurityContextHolder.getContext().setAuthentication(authToken);
-
-        filterChain.doFilter(request, response);
-    }
-}
