--- src/main/java/com/example/ddingsroom/repository/RefreshRepository.java
@@ -1,14 +1,0 @@
-package com.example.ddingsroom.repository;
-
-import com.example.ddingsroom.entity.RefreshEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-@Repository
-public interface RefreshRepository extends JpaRepository<RefreshEntity, Long> {
-    Boolean existsByRefresh(String refresh);
-
-    @Transactional
-    void deleteByRefresh(String refresh);
-}
