--- src/main/java/com/example/ddingsroom/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.example.ddingsroom.repository;
-
-import com.example.ddingsroom.entity.UserEntity;
-import com.example.ddingsroom.entity.VerificationCode;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-    Boolean existsByUsername(String username);
-    UserEntity findByUsername(String username);
-    Optional<UserEntity> findByEmail(String email);
-    Boolean existsByEmail(String email);
-}
