+++ src/main/java/com/example/ddingsroom/reservation/dto/BaseResponseDTO.java
@@ -1,0 +1,12 @@
+package com.example.ddingsroom.reservation.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class BaseResponseDTO {
+    private String message;
+}
