+++ src/main/java/com/example/ddingsroom/reservation/dto/ReservationCancelRequestDTO.java
@@ -1,0 +1,9 @@
+package com.example.ddingsroom.reservation.dto;
+
+import lombok.Data;
+
+@Data
+public class ReservationCancelRequestDTO {
+    private int userId;
+    private int reservationId;
+}
