+++ src/main/java/com/example/ddingsroom/reservation/dto/ReservationRequestDTO.java
@@ -1,0 +1,13 @@
+package com.example.ddingsroom.reservation.dto;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class ReservationRequestDTO {
+    private int userId;
+    private int roomId;
+    private LocalDateTime reservationStartTime;
+    private LocalDateTime reservationEndTime;
+}
