+++ src/main/java/com/example/ddingsroom/reservation/entity/ReservationEntity.java
@@ -1,0 +1,41 @@
+package com.example.ddingsroom.reservation.entity;
+
+import com.example.ddingsroom.user.entity.UserEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "reservation")
+@Data
+@Setter
+@Getter
+public class ReservationEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    
+    @ManyToOne
+    @JoinColumn(name = "room_id")
+    private RoomEntity room;
+    
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private UserEntity user;
+    
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+    private String status;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+}
