+++ src/main/java/com/example/ddingsroom/reservation/entity/RoomEntity.java
@@ -1,0 +1,23 @@
+package com.example.ddingsroom.reservation.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "room")
+@Data
+@Setter
+@Getter
+public class RoomEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    private String roomName;
+    private String roomStatus;
+}
