+++ src/main/java/com/example/ddingsroom/reservation/repository/RoomRepository.java
@@ -1,0 +1,9 @@
+package com.example.ddingsroom.reservation.repository;
+
+import com.example.ddingsroom.reservation.entity.RoomEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoomRepository extends JpaRepository<RoomEntity, Integer> {
+}
