--- src/main/java/com/example/ddingsroom/service/CustomUserDetailsService.java
@@ -1,27 +1,0 @@
-package com.example.ddingsroom.service;
-
-import com.example.ddingsroom.dto.CustomUserDetails;
-import com.example.ddingsroom.entity.UserEntity;
-import com.example.ddingsroom.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-    private final UserRepository userRepository;
-    public CustomUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        //DB에서 조회
-        UserEntity userData = userRepository.findByUsername(username);
-        if (userData != null) {
-            //UserDetails에 담아서 return하면 AutneticationManager가 검증 함
-            return new CustomUserDetails(userData);
-        }
-        return null;
-    }
-}
