+++ src/main/java/com/example/ddingsroom/user/controller/AdminController.java
@@ -1,0 +1,12 @@
+package com.example.ddingsroom.user.controller;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+@ResponseBody
+@Controller
+public class AdminController {
+    @GetMapping("/admin")
+    public String adminP() {
+        return "admin Controller";
+    }
+}
