+++ src/main/java/com/example/ddingsroom/user/dto/CodeSendDTO.java
@@ -1,0 +1,12 @@
+package com.example.ddingsroom.user.dto;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CodeSendDTO {
+    private String email;
+}
