+++ src/main/java/com/example/ddingsroom/user/dto/CodeVerifyDTO.java
@@ -1,0 +1,13 @@
+package com.example.ddingsroom.user.dto;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CodeVerifyDTO {
+    private String email;
+    private String code;
+}
