+++ src/main/java/com/example/ddingsroom/user/dto/ModifyPasswordDTO.java
@@ -1,0 +1,15 @@
+package com.example.ddingsroom.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ModifyPasswordDTO {
+    private String email;
+    private String password;
+}
