+++ src/main/java/com/example/ddingsroom/user/dto/SignUpDTO.java
@@ -1,0 +1,17 @@
+package com.example.ddingsroom.user.dto;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class SignUpDTO {
+    private String email;
+    private String username;
+    private String password;
+    private String age;
+    private String studentNumber;
+}
